backend/requirements.txt
openai
fastapi[all]
fastapi-cors
uvicorn
pydantic
langchain
langchain-pinecone
langchain-openai
langchain-community
pinecone
pandas
numpy
python-dotenv
PyMuPDF
json5